+++ pycognaize/tests/test_document/snapshot/test_exclude_options.py
@@ -1,0 +1,33 @@
+from unittest import TestCase
+
+from pycognaize import Snapshot
+
+
+class ExcludeOptionsTestCase(TestCase):
+
+    def test_should_exclude_image_folder_when_images_are_excluded(self):
+        res = Snapshot._get_exclude_patterns(True, False, False)
+
+        assert len(res) == 1
+        assert '*/images/*.jpeg' in res
+
+    def test_should_exclude_data_folder_when_ocr_is_excluded(self):
+        res = Snapshot._get_exclude_patterns(False, True, False)
+
+        assert len(res) == 1
+        assert '*/data/*.json' in res
+
+    def test_should_exclude_pdf_files_when_pdf_is_excluded(self):
+        res = Snapshot._get_exclude_patterns(False, False, True)
+
+        assert len(res) == 1
+        assert '*.pdf' in res
+
+    def test_should_include_all_options_when_all_are_excluded(self):
+        res = Snapshot._get_exclude_patterns(True, True, True)
+
+        assert len(res) == 3
+        assert '*.pdf' in res
+        assert '*/data/*.json' in res
+        assert '*/images/*.jpeg' in res
+